move CustomizationFieldDeleter to Update ns
